split advancedDiff into multiple modules
